Issues : MEJ-1, MEJ-3, MEJ-16
Several classes have been modified to be singleton.

The database connection are now made using the commons dbcp pool of connection. No more synchronization is needed to access the database. The connection is now is the ControlleurJeu. We now need to create a Database object by passing the connection to the constructor.

The rooms(and all the thing associated with it : objects, shops, rules) are now loaded from the database. Each time a player log in, a query is made to see if there is some new rooms to load).