Issue : MEJ-23, MEJ-32, MEJ-33
Start of the ER model using an eclipse database modeling tool : Clay (http://www.azzurri.jp/en/software/clay/index.jsp).

Rename of all tables and attributes to be in English.

There is now a level table (scholastic level). This level table is linked to the question_detail table to be able to have unlimited numbers of level not just secondary 1 to 5, etc...