Issue : MEJ-27
The server is now accepting policy request and sending back informations to the flash client. Added the binding to an asdress that can be different from the localhost to be able to run the server on the port 80 on a different interface but on same machine as the apache server.